src/backend/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Data processing
pandas==2.1.3
numpy==1.25.2

# StatsBomb data
statsbombpy==1.9.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
aiosqlite==0.19.0

# Caching
redis==5.0.1

# Environment variables
python-dotenv==1.0.0

# Static file serving
aiofiles==23.2.1